--- a/Cargo.toml
+++ b/Cargo.toml
@@ -50,12 +50,15 @@ unicode-segmentation = "1.11.0"
 boa_engine = "0.17.3"
 mime = "0.3.17"
 bytes = "1.5.0"
+flame = { version = "0.2.2", optional = true }
+flamer = { version = "0.5.0", optional = true }
 
 [dev-dependencies]
 tokio = { version = "1.36.0", features = ["full"] }
 
 [features]
 default = ["search", "live", "default-tls"]
+performance_analysis = ["flame", "flamer"]
 live = ["tokio/time", "tokio/process"]
 blocking = ["tokio/rt", "tokio/rt-multi-thread"]
 search = []
@@ -63,3 +66,7 @@ ffmpeg = ["tokio/process"]
 default-tls = ["reqwest/default-tls"]
 native-tls = ["reqwest/native-tls"]
 rustls-tls = ["reqwest/rustls-tls"]
+
+[[example]]
+name = "multiple_downloads"
+required-features = ["performance_analysis"]