--- a/cli/src/utils/mod.rs
+++ b/cli/src/utils/mod.rs
@@ -1,2 +1,1 @@
 pub mod result_serializer;
-