--- a/src/lib.rs
+++ b/src/lib.rs
@@ -6,8 +6,8 @@
 #[cfg(feature = "performance_analysis")]
 pub extern crate flame;
 #[cfg(feature = "performance_analysis")]
-#[macro_use] extern crate flamer;
-
+#[macro_use]
+extern crate flamer;
 
 mod info;
 mod info_extras;