--- a/src/utils.rs
+++ b/src/utils.rs
@@ -1,4 +1,5 @@
-use boa_engine::Context;
+use boa_engine::optimizer::OptimizerOptions;
+use boa_engine::{Context, JsValue, Source};
 use bytes::Bytes;
 use once_cell::sync::Lazy;
 use rand::Rng;
@@ -627,26 +628,35 @@ pub fn set_download_url(
         if let Some(&result) = n_transfrom_cache.get(n_transform_value).as_ref() {
             n_transform_result = result.clone();
         } else {
-            let mut context = boa_engine::Context::default();
-            let n_transform_script = context.eval(boa_engine::Source::from_bytes(
-                n_transform_script_string.1.as_str(),
-            ));
-
-            if n_transform_script.is_err() {
-                return url.to_string();
+            #[cfg_attr(feature = "performance_analysis", flamer::flame)]
+            // Caching this would be great (~2ms x 2 gain/req on Ryzen 9 5950XT) but is quite hard because of the !Send nature of boa
+            fn create_transform_script(script: &str) -> Context<'_> {
+                let mut context = boa_engine::Context::default();
+                context.eval(parse_source(
+                    script,
+                )).unwrap();
+                context
             }
 
-            let result = context.eval(boa_engine::Source::from_bytes(&format!(
-                r#"{func_name}("{args}")"#,
-                func_name = n_transform_script_string.0.as_str(),
-                args = n_transform_value
-            )));
-
-            if result.is_err() {
-                return url.to_string();
+            #[cfg_attr(feature = "performance_analysis", flamer::flame)]
+            fn parse_source<'a>(script: &'a str) -> Source<&'a [u8]> {
+                boa_engine::Source::from_bytes(script)
             }
 
-            let is_result_string = result.as_ref().unwrap().as_string();
+            #[cfg_attr(feature = "performance_analysis", flamer::flame)]
+            // Optimizing the script would be great (~20ms x 2 gain/req on Ryzen 9 5950XT) but quite some work on boa side
+            // This is where most of the time is spent
+            fn execute_transform_script(context: &mut Context, func_name: &str, n_transform_value: &str) -> JsValue{
+                context.eval(parse_source(&format!(
+                    r#"{func_name}("{n_transform_value}")"#,
+                ))).unwrap()
+            }
+
+            let mut context = create_transform_script(n_transform_script_string.1.as_str());
+
+            let is_result_string = execute_transform_script(&mut context, n_transform_script_string.0.as_str(), n_transform_value);
+
+            let is_result_string = is_result_string.as_string();
 
             if is_result_string.is_none() {
                 return url.to_string();